+++ CHANGES.md
@@ -1,3 +1,9 @@
 
+Version 0.2:
+ - Add doc strings to generated types
+ - Detect illegal names during package name conversion.
+ - Allow URIs to be truncated during package name conversion.
+ - Handle "." in class names
+
 Version 0.1:
  - Initial release